requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
scikit-learn
numpy